Split out the tests a bit more granularly and test more of the base scanner
